Fixed a somewhat obscure bug (refs issue_11).
In order to avoid crashing phpsh on fatal errors, a trick was added that would fork PHP on every
command.  This trick relies on `__init__.py` to keep track of PIDs.  Unfortunately, on some
systems, namely Ubuntu, `subprocess.Popen()` reports an incorrect PID for the initial startup of
`phpsh.php`.

The reason for this has to do with the way that `subprocess.Popen()` executes commands.  By
default, commands are executed using the Bourne shell (or a Bourne shell replacement which is
more likely the case) by issuing the following command: `/bin/sh -c command`.  The problem with
this is that some Bourne shell replacements handle this differently.  For example, in bash this
command will create a single process with a single PID.  In dash (the default Bourne shell
replacement for Ubuntu) this command will create two separate processes and report back to
Python that the PID is actually for the `/bin/sh -c` command.

To fix this inconsistency, we must force Python to use the same shell regardless of the operating
system.  I've chosen bash since it's the more ubiquitous Bourne shell replacement and also
produces the more desirable results.